\name{graphics}
\alias{graphics}
\title{
  Open a Windows Device
}
\description{
  Open a windows device in portrait or landscape mode.
}
\usage{
graphics(view = "portrait")
}
\arguments{
  \item{view}{if \code{"portrait"}, set width = 8.5 in and height = 11 in \cr
    if \code{"landscape"}, set width = 11 in and height = 8.5 in }
}
\seealso{
  \code{\link{resetGraph}}
}
\keyword{device}

