\name{plotLikes}
\alias{plotLikes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Plot Likelihood Profiles
}
\description{
  Plot likelihood profiles from files generated by \code{'repeatMPD'}.
}
\usage{
plotLikes(M=seq(0.02, 0.15, 0.01), A=c(40, 45, 50), R0=NULL,
   prefix="WWR-MMM-", znam="C@A_Commercial", zpos=1, 
   xlim=NULL, ylim=NULL, pnam=c("R0"), ppos=1:3, pxax="M", 
   outnam="LLcontours", type="contour", 
   png=FALSE, pngres=400, PIN=c(8, 8))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{M}{\code{numeric} -- vector of natural mortality values}
  \item{A}{\code{numeric} -- vector of age plus-classes}
  \item{R0}{\code{numeric} -- vector of time-zero (equilibrium) recruitment values}
  \item{prefix}{\code{character} -- prefix for output names, usually species code
    and capitals designating the varying factor in the likelihood analysis}
  \item{znam}{\code{character} -- Awatea name of variable in outpult likellihood file}
  \item{zpos}{\code{numeric} -- position (fleet or survey) for \code{'znam'}}
  \item{xlim}{\code{numeric} -- range of the x-axis}
  \item{ylim}{\code{numeric} -- range of the y-axis}
  \item{pnam}{\code{character} -- parameter name}
  \item{ppos}{\code{numeric} -- parameter position?}
  \item{pxax}{\code{character} -- parameter along x-axis}
  \item{outnam}{\code{character} -- output name}
  \item{type}{\code{character} -- type of plot (contour or lines)}
  \item{png}{\code{logical} -- if \code{TRUE}, send the figure to a png file}
  \item{pngres}{\code{numeric} -- resolution (pixels/inch) for the png file}
  \item{PIN}{\code{numeric} -- dimensions (plot size) in inches}
}
\details{
  Depends on output from running the function \code{'repeatMPD'}.
}
\value{
  Returns nothing in particular; plots likelihood profiles.
}
\author{
  \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{locus opus}: Regional Headquarters (RHQ), Vancouver BC\cr
  Last modified \code{Rd: 2023-07-04}
}
\seealso{
\pkg{PBSawatea}:
\code{\link[PBSawatea]{repeatMPD}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
\keyword{hplot}
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
\concept{likelihood profiles}
% Use only one concept per line.
