\name{plotTraj}
\alias{plotTraj}
\title{
  Plot Median Trajectory
}
\description{
  Plot median trajectories from various MCMC runs in one figure.
}
\usage{
plotTraj(dat, index, traj="B", y0=FALSE, lab.stock, 
   col=c("black","green4","blue","red","purple","orange"), 
   png=FALSE, pngres=400, PIN=c(8,8), lang=c("e","f"), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dat}{\code{list} -- an object (currently called \code{stock}) generated by
    running Sweave code to compile a Results Appendix. The object \code{stock} comprises
    one or more stocks, identified at the first level in the list's hierarchy.}
  \item{index}{\code{numeric} -- vector of indices pointing to runs that a user wishes to compare.
    For example, the index typically points to a base case and a number of sensitivity runs.}
  \item{traj}{\code{character} -- vector of trajectory names (e.g., \code{B} = spawning biomass,
    \code{Bt/B0} = spawning biomass depletion, etc.); options are limited.}
  \item{y0}{\code{logical} -- if \code{TRUE}, start the y-axis at 0.}
  \item{lab.stock}{\code{character} -- name of stock to appear in the name of the
    \code{png} output file, if \code{png=TRUE}.}
  \item{col}{\code{character} -- names of colours for each trajectory.}
  \item{png}{\code{logical} -- if \code{TRUE}, send the figure output to a \code{png} device.}
  \item{pngres}{\code{numeric} -- resolution (pixels/inch) for \code{png} file.}
  \item{PIN}{\code{numeric} -- vector of width and height (inches) for \code{png} file.}
  \item{lang}{\code{character} -- a vector of letters that denote the language for output:
    currently only \code{"e"} (english) and \code{"f"} (french).}
  \item{\dots}{\code{list} -- additional arguments for \code{lines} (\code{lty} only) and \code{legend}.}
}
\details{
  Trajectories are currently limited to objects in the \code{currentMCMC} object 
  (\code{c("B","VB","R","U")}) and spawning biomass depletion (\code{Bt/B0}).
}
\value{
  Assigns a trajectory \code{array} object to the global environment.
}
\author{
  \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{locus: Institute of Ocean Sciences (IOS), Sidney BC}\cr
  Last modified \code{Rd: 2018-07-26}
}
\note{
  This function is geared for use after a number of Awatea model runs have been completed.
  The steps for creating one model run generally involve five stages -- (i) reweighting
  abundance & composition data, (ii) MPD fit, (iii) MCMC fits, (iv) MSY calculations, and
  (v) projections. After a number of model runs have been generated, the results can be
  gathered and compared.
}
\seealso{
\code{\link[PBSawatea]{runADMB}},
\code{\link[PBSawatea]{runMPD}},
\code{\link[PBSawatea]{runMCMC}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{hplot}% use one of  RShowDoc("KEYWORDS")
