\name{plt.expRate}
\alias{plt.expRate}
\title{
  Plot Exploitation Rate
}
\description{
  Plot exploitation rate against year.
}
\usage{
plt.expRate(obj, yLim=c(0,0.5), xLim=c(1954,2005))
}
\arguments{
  \item{obj}{an object from \code{load.allResFiles}.}
  \item{yLim}{limits of the y-axis.}
  \item{xLim}{limits of the x-axis.}
}
\details{
  Simple points and lines plot.
}
\author{
  Andrew Edwards, Pacific Biological Station, Fisheries and Oceans Canada, Nanaimo BC
}
\seealso{
  \code{\link{plt.idx}}
}
\keyword{hplot}
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line

