\name{plt.numR}
\alias{plt.numR}
\title{
  Plot Numbers at Age at Equilibrium
}
\description{
  Plot numbers at age at equilibrium. Plot recruitment (age 1).
}
\usage{
plt.numR(obj, minYr = NULL)
}
\arguments{
  \item{obj}{an Awatea results object from \code{load.allResFiles}.}
  \item{minYr}{minimum year to display in plot.}
}
\author{
  Andrew Edwards, Pacific Biological Station, Fisheries and Oceans Canada, Nanaimo BC
}
\note{
  Not sure what this plot is used for (RH).
}
\seealso{
  \code{\link{plotRmcmcPOP}}
}
\keyword{hplot}
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
